INT-2828 Remove Class Tangle in JPA
